CODE/requirements.txt
torch==2.6.0
torchvision==0.21.0
projectaria-tools[all]==2.0.0
pyyaml
matplotlib
packaging
h5py
ipython
rich
wandb
hydra-core
hydra-submitit-launcher==1.2.0
black
gpustat
pynvml
termcolor
pyquaternion
rospkg
einops
av==12.0.0
opencv-python
dm-control==1.0.8
mink==0.0.13
mujoco==3.4.0
mujoco-py==2.1.2.14
submitit
arm_pytorch_utilities
pytorch-kinematics
lightning
positional-encodings[pytorch]
argparse
pandas
overrides
zarr==3.1.5
tabulate
transformers==4.57.3
timm
boto3
cloudpathlib
sqlalchemy
psycopg
ray
geomloss
tslearn
sqlalchemy
psycopg[binary]
boto3
typing_extensions
pyarrow
simplejpeg
prettytable
datasets==4.0.0
s5cmd
mediapy
pytest
pre-commit
ruff
scaleapi
openai